+++ src/main/java/com/example/willhero/MediumOrc.java
@@ -1,8 +1,8 @@
 package com.example.willhero;
 
 public class MediumOrc extends Orc{
-    MediumOrc(double x, double y,double length,double breadth,double xSpeed,
-              double ySpeed, String color, int coinsOnKill, int health,String imageURL){
+    MediumOrc(double x, double y,double length,double breadth,int xSpeed,
+              int ySpeed, String color, int coinsOnKill, int health,String imageURL){
         super(x,y,length,breadth,xSpeed,ySpeed,color,coinsOnKill,health,imageURL);
     }
 }
