+++ src/main/java/com/example/willhero/SmallOrc.java
@@ -1,9 +1,9 @@
 package com.example.willhero;
 
 public class SmallOrc extends Orc{
-    SmallOrc(double x, double y,double length, double breadth,double xSpeed,
-             double ySpeed, String color, int coinsOnKill, int health,String imageURL){
+    SmallOrc(double x, double y,double length, double breadth,int xSpeed,
+             int ySpeed, String color, int coinsOnKill, int health,String imageURL){
         super(x,y,length,breadth,xSpeed,ySpeed,color,coinsOnKill,health,imageURL);
-
     }
+
 }
